fix wrong module import
